He is in charge of controlling the game.
Create a composition with the BoardManager class to call its Setup() method.

It has a Singleton design pattern to create an instance of it in a static way to call it from any Script.

It contains a special method so that it always exists in the scene if it is deleted.